refactor(mcp): reorganize UI components into ui package #371
Move MCP UI components (ToolDetailPanel, LlmConfig, ToolListPanel, ResultPanel) from editor to ui package for better code organization. Extracted McpLlmConfig to separate file and updated
