+++ core/src/main/kotlin/cc/unitmesh/devti/mcp/ui/McpResultPanel.kt
@@ -1,4 +1,4 @@
-package cc.unitmesh.devti.mcp.editor
+package cc.unitmesh.devti.mcp.ui
 
 import cc.unitmesh.devti.util.parser.CodeFence
 import com.intellij.ui.JBColor
